Backend/service_logs/requirements.txt
djangorestframework==3.15.2
environ==1.0
drf-yasg==1.21.8
django-environ==0.11.2
psycopg==3.2.3
celery==5.4.0
django-celery-results==2.5.1
kombu==5.4.2
django-oauth-toolkit==3.0.1
django-cors-headers==4.6.0